+++ CHANGELOG.md
@@ -18,3 +18,29 @@
 ## 0.0.5
 
 * Removed mock implementations
+
+## 0.0.6
+
+* Improved Packages 
+
+## 0.0.7
+
+* Fixed issue where Android would not build
+
+## 0.0.8
+
+* Fixed issue where Android would not build
+
+
+## 0.1.0
+
+* Fixed issue where number is not a number
+
+
+## 0.1.1
+
+* Fixed issue where number is not a number
+
+## 0.1.2
+
+* Removed debug logs